refactor(#1780): extract fork lifecycle into operational_manifest/fork.rs to satisfy 2000-line file-contract guard
